Titles with multiple names for arguments.
This change adds the posibility to add multiple names for arguments,
e.g.  in a `git remote add` command the arguments can be `name` and
`url`, therefore in the help message it's going to render `<name>` and
`<url>` as separate arguments. A further improvement would be to handle
tpyes and arity for arguments too.